[refactor] move conversion from src level to spdx3
