+++ src/spdx3/conversion/message.py
