+++ src/spdx3/conversion/snippet_conversion.py
@@ -8,7 +8,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from conversion.message import print_missing_conversion
+from spdx3.conversion.message import print_missing_conversion
 from spdx3.model.software.snippet import Snippet
 
 from spdx3.model.creation_information import CreationInformation
